[feat] 트레이너 상세 프로필 구현/사진 관리 개선
- TrainerController.java 삭제
- index.html 삭제
- MainController.java 수정: 메인페이지(트레이너 목록)
- TrainerService.java 수정
- trainers.html 수정
- trainers.css 추가
- trainer-detail.html 추가
- trainer-detail.css 추가
