+++ src/main/resources/static/css/trainers.css
@@ -1,0 +1,40 @@
+@charset "UTF-8";
+
+/* 트레이너 프로필 목록(메인페이지) */
+.trainer-container {
+    display: grid;
+    grid-template-columns: repeat(3, 1fr);
+    gap: 20px;
+    padding: 20px;
+}
+
+.trainer-card {
+    border: 1px solid #ddd;
+    border-radius: 10px;
+    padding: 15px;
+    text-align: center;
+    background-color: #f9f9f9;
+    box-shadow: 2px 2px 10px rgba(0, 0, 0, 0.1);
+}
+
+.trainer-thumbnail {
+    width: 100%;
+    height: 180px;
+    object-fit: cover; /* 사진 크기 조정 */
+    border-radius: 10px;
+}
+
+.trainer-card a {
+    display: inline-block;
+    padding: 8px 15px;
+    background-color: #007bff;
+    color: white;
+    text-decoration: none;
+    border-radius: 5px;
+    font-weight: bold;
+    text-align: center;
+}
+
+.trainer-card a:hover {
+    background-color: #0056b3;
+}
