+++ src/main/resources/templates/trainers.html
@@ -1,27 +1,26 @@
 <!DOCTYPE html>
-<html lang="ko" xmlns:th="http://www.thymeleaf.org">
+<html xmlns:th="http://www.thymeleaf.org">
 <head>
     <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>트레이너 목록</title>
-    <style>
-        body { font-family: Arial, sans-serif; margin: 20px; }
-        .trainer-card { border: 1px solid #ddd; padding: 10px; margin: 10px; border-radius: 8px; }
-        img { max-width: 150px; display: block; margin-bottom: 10px; }
-    </style>
+    
+    <!-- Bootstrap 포함 -->
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css">
+    
+    <!-- 내 CSS -->
+    <link rel="stylesheet" th:href="@{/css/trainers.css}">
 </head>
+
 <body>
     <h1>트레이너 목록</h1>
     
-    <div th:each="trainer : ${trainers}" class="trainer-card">
-        <img th:src="${trainer.photo}" th:alt="${trainer.name}">
-        <h3 th:text="${trainer.name}"></h3>
-        <p>전문 분야: <span th:text="${trainer.specialization}"></span></p>
-        <p>경력: <span th:text="${trainer.experience}"></span>년</p>
-        <p>가격: <span th:text="${trainer.fee}"></span>원</p>
-        <p th:text="${trainer.bio}"></p>
+    <div class="trainer-container">
+        <div class="trainer-card" th:each="trainer : ${trainers}">
+        	<img th:src="${trainer.photos[0]}" alt="Trainer Image" class="trainer-thumbnail">
+            <h2 th:text="${trainer.name}">트레이너 이름</h2>
+			<p th:text="${trainer.specialization}"></p>
+			<a th:href="@{/trainer/{id}(id=${trainer.trainerId})}" class="btn btn-primary">View Details</a>
+        </div>
     </div>
-
-    <a href="/"><button>홈으로</button></a>
 </body>
 </html>
